Require in Accounts/Events Schema
Account and Events Schema now have requirements to ensure that requests with empty fields will not be accepted. This is already handled by the client, but the extra layer of security will be useful, especially if attempting manual insertions.

Also removed to auth server code (really not that important). Was hoping to add a separate server to provide access and refresh tokens, but there is honestly just not enough time for that